backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
google-genai
pydantic
python-dotenv
httpx
pytest
pytest-asyncio
